@@ -12,18 +12,17 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^7.2.15",
+    "@angular/animations": "^8.2.10",
     "@angular/cdk": "^7.3.7",
-    "@angular/common": "^7.2.15",
-    "@angular/compiler": "^7.2.15",
-    "@angular/core": "^7.2.15",
+    "@angular/common": "^8.2.10",
+    "@angular/compiler": "^8.2.10",
+    "@angular/core": "^8.2.10",
     "@angular/fire": "^5.2.1",
-    "@angular/forms": "^7.2.15",
-    "@angular/http": "^7.2.15",
+    "@angular/forms": "^8.2.10",
     "@angular/material": "^7.3.7",
-    "@angular/platform-browser": "^7.2.15",
-    "@angular/platform-browser-dynamic": "^7.2.15",
-    "@angular/router": "^7.2.15",
+    "@angular/platform-browser": "^8.2.10",
+    "@angular/platform-browser-dynamic": "^8.2.10",
+    "@angular/router": "^8.2.10",
     "@fortawesome/angular-fontawesome": "^0.3.0",
     "@fortawesome/fontawesome-svg-core": "^1.2.25",
     "@fortawesome/free-solid-svg-icons": "^5.11.2",
@@ -38,22 +37,22 @@
     "ngx-contextmenu": "^5.2.0",
     "ngx-gallery": "^5.10.0",
     "replace-in-file": "^3.4.4",
-    "rxjs": "~6.3.3",
+    "rxjs": "~6.5.3",
     "sudoku": "0.0.2",
     "tslib": "^1.10.0",
     "webpack-dev-server": "^3.8.2",
-    "zone.js": "^0.8.29"
+    "zone.js": "~0.9.1"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~0.7.0",
-    "@angular/cli": "^7.3.9",
-    "@angular/compiler-cli": "^7.2.15",
-    "@angular/language-service": "^7.2.15",
+    "@angular-devkit/build-angular": "~0.803.9",
+    "@angular/cli": "^8.3.9",
+    "@angular/compiler-cli": "^8.2.10",
+    "@angular/language-service": "^8.2.10",
     "@types/jasmine": "^2.8.16",
     "@types/jasminewd2": "^2.0.8",
     "@types/node": "~8.9.4",
     "angular-cli-ghpages": "^0.5.3",
-    "codelyzer": "~4.3.0",
+    "codelyzer": "^5.0.1",
     "gh-pages": "^2.1.1",
     "jasmine-core": "~2.99.1",
     "jasmine-spec-reporter": "~4.2.1",
@@ -65,6 +64,6 @@
     "protractor": "^5.4.2",
     "ts-node": "~7.0.0",
     "tslint": "~5.11.0",
-    "typescript": "~3.1.6"
+    "typescript": "~3.5.3"
   }
 }
